Рефакторинг ProfileViewController. Перенос показа алерта в ProfileViewPresenter
